backend/requirements.txt
fastapi
uvicorn
yfinance
transformers
torch
pandas
sqlmodel
python-dotenv
scikit-learn
httpx
httpx
